Macadamian Help
- Edited the Guest home page so that it actually closes without all of
the other buttons in it.
- Edited the GuestSinglePlayerGamePage to swap between pages the way we
were thinking, it doesn't work yet, however we talked about how we
should do it. Use the divs to swap the pages, only place the page
content, all css and js files in one a global file, using a hide() to
display and Hide the different pages.